feat(admin): enhance admin dashboard with new components for settings and ingest targets management
