+++ frontend/vite.config.ts
@@ -1,7 +1,7 @@
 import { defineConfig, loadEnv } from 'vite';
 import deno from '@deno/vite-plugin';
 import react from '@vitejs/plugin-react-swc';
-import { TanStackRouterVite } from '@tanstack/router-plugin/vite';
+import { tanstackRouter } from '@tanstack/router-plugin/vite';
 import { resolve } from 'node:path';
 import process from 'node:process';
 
@@ -11,7 +11,9 @@ export default defineConfig(({ mode }: { mode: string }) => {
     const env = loadEnv(mode, process.cwd(), '');
 
     return {
-        plugins: [deno(), react(), TanStackRouterVite()],
+        plugins: [deno(), react(), tanstackRouter({
+            addExtensions: true
+        })],
 
         resolve: {
             alias: {
